+++ test/linkify_test.exs
@@ -386,6 +386,15 @@ defmodule LinkifyTest do
              ) == expected
     end
 
+    test "mentions with symbols before them" do
+      text = "@@example hey! >@@test@example.com"
+
+      expected =
+        "@<a href=\"/users/example\">@example</a> hey! >@<a href=\"/users/test@example.com\">@test@example.com</a>"
+
+      assert Linkify.link(text, mention: true, mention_prefix: "/users/") == expected
+    end
+
     test "invalid mentions" do
       text = "hey user@example"