+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Hashtags followed by HTML tags "a", "code" and "pre" were not detected
 - Incorrect parsing of HTML links inside HTML tags
 - Punctuation marks in the end of urls were included in the html links
+- Incorrect parsing of mentions with symbols before them
 
 ## 0.2.0 - 2020-07-21